Backend/requirements.txt
flask
pymysql
sqlalchemy
jsonify
python-dotenv
gunicorn
gevent
flask_cors
redis
python-redis-lock